requirements.txt
numpy>=1.19.0
pytest>=6.0.0
flask>=2.0.0
pillow>=8.0.0
tqdm>=4.65.0
torch>=1.8.0
torchvision>=0.9.0 